perf(quest): store key envelopes in bf16
An envelope entry is the min or max of a set of bf16 keys, so it is already
exactly a bf16 value; fp32 storage only ever stored the same number in twice
the space. `envelope_dot` widens bf16->fp32 exactly and accumulates in f32, so
every arithmetic result is bit-identical -- `test_envelope_dot`'s 13 checks,
including the exact merge-equals-recompute equality and the golden vector, pass
unchanged with no tolerance loosened.

Directed rounding (`_rd` for min, `_ru` for max) is a no-op on every value the
current code stores. It is there so a future caller merging a value that did
not originate as a bf16 key cannot round a bound inwards, which would drop a
live key from the score and evict a page that should have been kept.

Min/max atomics move to `atomicCAS` on `unsigned short`; bf16 is the top 16 bits
of fp32, so the sign-magnitude ordering argument carries over verbatim.

Envelopes now cost 6.25% of a KV page rather than 12.5%. Measured on an L40S:
22084 pages with the switch off, 20785 with it on (ratio 1.0625 exactly). Under
fp32 the same device held 19630 -- the narrowing hands back 1155 pages, 18480
tokens of KV capacity.

Documented as section 16.

+++ driver/cuda/src/pipeline/grouped_runtime.cuh
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cuda_bf16.h>
+
 #include <algorithm>
 #include <array>
 #include <cstddef>
@@ -62,8 +64,8 @@ struct GroupedLanePageMask {
 };
 
 struct GroupedLaneEnvelopeDevice {
-    const float* env_min;
-    const float* env_max;
+    const __nv_bfloat16* env_min;
+    const __nv_bfloat16* env_max;
     const float* query;
     // UNSLICED fire page list, plus the device CSR that slices it. See the
     // comment on `GroupedLaneEnvelope` for why the host must not pre-slice.
@@ -94,8 +96,8 @@ struct GroupedLaneEnvelopeDevice {
 // `page_indptr` instead, and the host contributes only the *declared slot
 // bound*, which is safe to over-estimate.
 struct GroupedLaneEnvelope {
-    const float* env_min = nullptr;   // [num_pages, num_kv_heads, head_dim]
-    const float* env_max = nullptr;
+    const __nv_bfloat16* env_min = nullptr;   // [num_pages, num_kv_heads, head_dim]
+    const __nv_bfloat16* env_max = nullptr;
     // This lane's query row to score with: the LAST token of the lane, which is
     // the token whose attention the mask will govern.
     const float* query = nullptr;
